requirements.txt
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.6
inflate64==1.0.0
multivolumefile==0.2.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.3.0
prettytable==3.10.0
psutil==5.9.8
py7zr==0.21.0
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.20.0
PyMySQL==1.1.0
pyppmd==1.1.0
pyTelegramBotAPI==4.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzstd==0.15.10
requests==2.31.0
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.29
telebot==0.0.5
texttable==1.7.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
